+++ backend/apid/actions/events_test.go
@@ -408,7 +408,7 @@ func TestEventUpdate(t *testing.T) {
 func TestEventCreate(t *testing.T) {
 	defaultCtx := testutil.NewContext(
 		testutil.ContextWithRules(
-			types.FixtureRuleWithPerms(types.RuleTypeEvent, types.RulePermCreate, types.RulePermUpdate),
+			types.FixtureRuleWithPerms(types.RuleTypeEvent, types.RulePermCreate),
 		),
 	)
 	wrongPermsCtx := testutil.NewContext(
@@ -437,11 +437,12 @@ func TestEventCreate(t *testing.T) {
 			expectedErr: false,
 		},
 		{
-			name:        "Already Exists",
-			ctx:         defaultCtx,
-			argument:    types.FixtureEvent("entity1", "check1"),
-			fetchResult: types.FixtureEvent("entity1", "check1"),
-			expectedErr: false,
+			name:            "Already Exists",
+			ctx:             defaultCtx,
+			argument:        types.FixtureEvent("entity1", "check1"),
+			fetchResult:     types.FixtureEvent("entity1", "check1"),
+			expectedErr:     true,
+			expectedErrCode: AlreadyExistsErr,
 		},
 		{
 			name:            "Store Err on Fetch",
@@ -492,6 +493,103 @@ func TestEventCreate(t *testing.T) {
 
 			// Exec Query
 			err := actions.Create(tc.ctx, *tc.argument)
+			if tc.expectedErr {
+				inferErr, ok := err.(Error)
+				if ok {
+					assert.Equal(tc.expectedErrCode, inferErr.Code)
+				} else {
+					assert.Error(err)
+					assert.FailNow("Given was not of type 'Error'")
+				}
+			} else {
+				assert.NoError(err)
+			}
+		})
+	}
+}
+
+func TestEventCreateOrReplace(t *testing.T) {
+	defaultCtx := testutil.NewContext(
+		testutil.ContextWithRules(
+			types.FixtureRuleWithPerms(
+				types.RuleTypeEvent,
+				types.RulePermCreate,
+				types.RulePermUpdate,
+			),
+		),
+	)
+	wrongPermsCtx := testutil.NewContext(
+		testutil.ContextWithRules(
+			types.FixtureRuleWithPerms(types.RuleTypeEvent, types.RulePermCreate),
+		),
+	)
+
+	badEvent := types.FixtureEvent("entity1", "check1")
+	badEvent.Check.Name = "!@#!#$@#^$%&$%&$&$%&%^*%&(%@###"
+
+	testCases := []struct {
+		name            string
+		ctx             context.Context
+		argument        *types.Event
+		fetchResult     *types.Event
+		fetchErr        error
+		busErr          error
+		expectedErr     bool
+		expectedErrCode ErrCode
+	}{
+		{
+			name:        "Created",
+			ctx:         defaultCtx,
+			argument:    types.FixtureEvent("entity1", "check1"),
+			expectedErr: false,
+		},
+		{
+			name:        "Already Exists",
+			ctx:         defaultCtx,
+			argument:    types.FixtureEvent("entity1", "check1"),
+			fetchResult: types.FixtureEvent("entity1", "check1"),
+		},
+		{
+			name:            "No Permission",
+			ctx:             wrongPermsCtx,
+			argument:        types.FixtureEvent("entity1", "check1"),
+			expectedErr:     true,
+			expectedErrCode: PermissionDenied,
+		},
+		{
+			name:            "Validation Error",
+			ctx:             defaultCtx,
+			argument:        badEvent,
+			expectedErr:     true,
+			expectedErrCode: InvalidArgument,
+		},
+		{
+			name:            "Message Bus Error",
+			ctx:             defaultCtx,
+			argument:        types.FixtureEvent("entity1", "check1"),
+			busErr:          errors.New("where's the wizard"),
+			expectedErr:     true,
+			expectedErrCode: InternalErr,
+		},
+	}
+
+	for _, tc := range testCases {
+		store := &mockstore.MockStore{}
+		bus := &mockbus.MockBus{}
+		actions := NewEventController(store, bus)
+
+		t.Run(tc.name, func(t *testing.T) {
+			assert := assert.New(t)
+
+			// Mock store methods
+			store.
+				On("GetEventByEntityCheck", mock.Anything, mock.Anything, mock.Anything).
+				Return(tc.fetchResult, tc.fetchErr)
+
+			bus.On("Publish", mock.Anything, mock.Anything).Return(tc.busErr)
+
+			// Exec Query
+			err := actions.CreateOrReplace(tc.ctx, *tc.argument)
 			if tc.expectedErr {
 				inferErr, ok := err.(Error)
 				if ok {