+++ backend/apid/graphql/environment.go
@@ -21,12 +21,16 @@ var _ schema.EnvironmentFieldResolvers = (*envImpl)(nil)
 
 type envImpl struct {
 	orgCtrl    actions.OrganizationsController
+	checksCtrl actions.CheckController
+	entityCtrl actions.EntityController
 	eventsCtrl actions.EventController
 }
 
-func newEnvImpl(store store.Store) *envImpl {
+func newEnvImpl(store store.Store, getter types.QueueGetter) *envImpl {
 	return &envImpl{
 		orgCtrl:    actions.NewOrganizationsController(store),
+		checksCtrl: actions.NewCheckController(store, getter),
+		entityCtrl: actions.NewEntityController(store),
 		eventsCtrl: actions.NewEventController(store, nil),
 	}
 }
@@ -55,6 +59,48 @@ func (r *envImpl) Organization(p graphql.ResolveParams) (interface{}, error) {
 	return handleControllerResults(org, err)
 }
 
+// Checks implements response to request for 'checks' field.
+func (r *envImpl) Checks(p schema.EnvironmentChecksFieldResolverParams) (interface{}, error) {
+	env := p.Source.(types.Environment)
+	ctx := types.SetContextFromResource(p.Context, &env)
+	records, err := r.checksCtrl.Query(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	// pagination
+	info := relay.NewArrayConnectionInfo(
+		0, len(records),
+		p.Args.First, p.Args.Last, p.Args.Before, p.Args.After,
+	)
+	edges := make([]*relay.Edge, info.End-info.Begin)
+	for i, r := range records[info.Begin:info.End] {
+		edges[i] = relay.NewArrayConnectionEdge(r, i)
+	}
+	return relay.NewArrayConnection(edges, info), nil
+}
+
+// Entities implements response to request for 'entities' field.
+func (r *envImpl) Entities(p schema.EnvironmentEntitiesFieldResolverParams) (interface{}, error) {
+	env := p.Source.(types.Environment)
+	ctx := types.SetContextFromResource(p.Context, &env)
+	records, err := r.entityCtrl.Query(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	info := relay.NewArrayConnectionInfo(
+		0, len(records),
+		p.Args.First, p.Args.Last, p.Args.Before, p.Args.After,
+	)
+
+	edges := make([]*relay.Edge, info.End-info.Begin)
+	for i, r := range records[info.Begin:info.End] {
+		edges[i] = relay.NewArrayConnectionEdge(r, i)
+	}
+	return relay.NewArrayConnection(edges, info), nil
+}
+
 // Events implements response to request for 'events' field.
 func (r *envImpl) Events(p schema.EnvironmentEventsFieldResolverParams) (interface{}, error) {
 	env := p.Source.(types.Environment)
@@ -64,6 +110,7 @@ func (r *envImpl) Events(p schema.EnvironmentEventsFieldResolverParams) (interfa
 		return nil, err
 	}
 
+	// apply filters
 	var filteredEvents []*types.Event
 	filter := p.Args.Filter
 	if len(filter) > 0 {
@@ -83,6 +130,7 @@ func (r *envImpl) Events(p schema.EnvironmentEventsFieldResolverParams) (interfa
 		filteredEvents = records
 	}
 
+	// sort records
 	if p.Args.OrderBy == schema.EventsListOrders.SEVERITY {
 		sort.Sort(types.EventsBySeverity(filteredEvents))
 	} else {
@@ -92,11 +140,11 @@ func (r *envImpl) Events(p schema.EnvironmentEventsFieldResolverParams) (interfa
 		))
 	}
 
+	// pagination
 	info := relay.NewArrayConnectionInfo(
 		0, len(filteredEvents),
 		p.Args.First, p.Args.Last, p.Args.Before, p.Args.After,
 	)
-
 	edges := make([]*relay.Edge, info.End-info.Begin)
 	for i, r := range filteredEvents[info.Begin:info.End] {
 		edges[i] = relay.NewArrayConnectionEdge(r, i)