+++ backend/apid/routers/checks.go
@@ -27,11 +27,11 @@ func NewChecksRouter(store store.Store, getter types.QueueGetter) *ChecksRouter 
 // Mount the ChecksRouter to a parent Router
 func (r *ChecksRouter) Mount(parent *mux.Router) {
 	routes := resourceRoute{router: parent, pathPrefix: "/checks"}
-	routes.index(r.list)
-	routes.show(r.find)
-	routes.create(r.create)
-	routes.update(r.update)
-	routes.destroy(r.destroy)
+	routes.getAll(r.list)
+	routes.get(r.find)
+	routes.post(r.create)
+	routes.del(r.destroy)
+	routes.put(r.createOrReplace)
 
 	// Custom
 	routes.path("{id}/hooks/{type}", r.addCheckHook).Methods(http.MethodPut)
@@ -66,13 +66,13 @@ func (r *ChecksRouter) create(req *http.Request) (interface{}, error) {
 	return cfg, err
 }
 
-func (r *ChecksRouter) update(req *http.Request) (interface{}, error) {
+func (r *ChecksRouter) createOrReplace(req *http.Request) (interface{}, error) {
 	cfg := types.CheckConfig{}
 	if err := unmarshalBody(req, &cfg); err != nil {
 		return nil, err
 	}
 
-	err := r.controller.Update(req.Context(), cfg)
+	err := r.controller.CreateOrReplace(req.Context(), cfg)
 	return cfg, err
 }