+++ backend/apid/routers/entities.go
@@ -7,7 +7,6 @@ import (
 	"github.com/gorilla/mux"
 	"github.com/sensu/sensu-go/backend/apid/actions"
 	"github.com/sensu/sensu-go/backend/store"
-	"github.com/sensu/sensu-go/types"
 )
 
 // EntitiesRouter handles requests for /entities
@@ -25,10 +24,9 @@ func NewEntitiesRouter(store store.EntityStore) *EntitiesRouter {
 // Mount the EntitiesRouter to a parent Router
 func (r *EntitiesRouter) Mount(parent *mux.Router) {
 	routes := resourceRoute{router: parent, pathPrefix: "/entities"}
-	routes.destroy(r.destroy)
-	routes.index(r.list)
-	routes.show(r.find)
-	routes.update(r.update)
+	routes.getAll(r.list)
+	routes.get(r.find)
+	routes.del(r.destroy)
 }
 
 func (r *EntitiesRouter) destroy(req *http.Request) (interface{}, error) {
@@ -54,14 +52,4 @@ func (r *EntitiesRouter) find(req *http.Request) (interface{}, error) {
 func (r *EntitiesRouter) list(req *http.Request) (interface{}, error) {
 	records, err := r.controller.Query(req.Context())
 	return records, err
-}
-
-func (r *EntitiesRouter) update(req *http.Request) (interface{}, error) {
-	entity := types.Entity{}
-	if err := unmarshalBody(req, &entity); err != nil {
-		return nil, err
-	}
-
-	err := r.controller.Update(req.Context(), entity)
-	return entity, err
 }