+++ backend/apid/routers/users.go
@@ -25,11 +25,11 @@ func NewUsersRouter(store store.Store) *UsersRouter {
 // Mount the UsersRouter to a parent Router
 func (r *UsersRouter) Mount(parent *mux.Router) {
 	routes := resourceRoute{router: parent, pathPrefix: "/rbac/users"}
-	routes.index(r.list)
-	routes.show(r.find)
-	routes.create(r.create)
-	routes.update(r.update)
-	routes.destroy(r.destroy)
+	routes.getAll(r.list)
+	routes.get(r.find)
+	routes.post(r.create)
+	routes.del(r.destroy)
+	routes.put(r.createOrReplace)
 
 	// Custom
 	routes.path("{id}/reinstate", r.reinstate).Methods(http.MethodPut)
@@ -72,20 +72,20 @@ func (r *UsersRouter) create(req *http.Request) (interface{}, error) {
 
 	err := r.controller.Create(req.Context(), cfg)
 
-	// Obfustace users password
+	// Hide user's password
 	cfg.Password = ""
 	return cfg, err
 }
 
-func (r *UsersRouter) update(req *http.Request) (interface{}, error) {
+func (r *UsersRouter) createOrReplace(req *http.Request) (interface{}, error) {
 	cfg := types.User{}
 	if err := unmarshalBody(req, &cfg); err != nil {
 		return nil, err
 	}
 
-	err := r.controller.Update(req.Context(), cfg)
+	err := r.controller.CreateOrReplace(req.Context(), cfg)
 
-	// Obfustace users password
+	// Hide user's password
 	cfg.Password = ""
 	return cfg, err
 }